updates to hip (was hip_bench) -- HipPrjn was returning DoSynCa = false which should have significantly imparied learning.. seems maybe better now but hard to tell.
