+++ examples/hip/def.params
