+++ examples/hip/def_memory.png
