+++ examples/hip/diff/1vs2_diffs_2.png
