+++ examples/hip/diff/1vs2_diffs_3.png
