+++ examples/hip/diff/1vs3_diffs_1.png
