+++ examples/hip/diff/1vs3_diffs_2.png
