+++ examples/hip/diff/2vs4_diffs_1.png
