+++ examples/hip/diff/2vs4_diffs_2.png
