+++ examples/hip/diff/README.md
