+++ examples/hip/diff/fig_alphacyc_diffs_1.png
