+++ examples/hip/diff/fig_alphacyc_diffs_2.png
