+++ examples/hip/diff/fig_alphacyc_diffs_3.png
