+++ examples/hip/diff/fig_netconfig_diffs.png
