+++ examples/hip/grunter_ror.py
