+++ examples/hip/jobs.archive
@@ -1,0 +1,1 @@
+$JobId,$Server,$SlurmId,$Status,$SlurmStat,$Submit,$Start,$End,$Args,$Message
