+++ examples/hip/orig_memory.png
