+++ examples/hip/orig_params.go
