--- examples/hip/params.go
@@ -1,97 +1,0 @@
-// File generated by params.SaveGoCode
-
-//go:build this_is_broken_we_should_fix_or_delete
-
-package main
-
-import "github.com/emer/emergent/params"
-
-var SavedParamsSets = params.Sets{
-	{Name: "Base", Desc: "these are the best params", Sheets: params.Sheets{
-		"Network": &params.Sheet{
-			{Sel: "Prjn", Desc: "keeping default params for generic prjns",
-				Params: params.Params{}},
-			{Sel: ".EcCa1Prjn", Desc: "encoder projections -- no norm, moment",
-				Params: params.Params{
-					"Prjn.Learn.LRate": "0.04",
-				}},
-			{Sel: ".HippoCHL", Desc: "hippo CHL projections -- no norm, moment, but YES wtbal = sig better",
-				Params: params.Params{
-					"Prjn.CHL.Hebb":    "0.05",
-					"Prjn.Learn.LRate": "0.4",
-				}},
-			{Sel: "#CA1ToECout", Desc: "extra strong from CA1 to ECout",
-				Params: params.Params{
-					"Prjn.PrjnScale.Abs": "4.0",
-				}},
-			{Sel: "#InputToECin", Desc: "one-to-one input to EC",
-				Params: params.Params{
-					"Prjn.Learn.Learn": "false",
-					"Prjn.WtInit.Mean": "0.8",
-					"Prjn.WtInit.Var":  "0.0",
-				}},
-			{Sel: "#ECoutToECin", Desc: "one-to-one out to in",
-				Params: params.Params{
-					"Prjn.Learn.Learn":   "false",
-					"Prjn.WtInit.Mean":   "0.9",
-					"Prjn.WtInit.Var":    "0.01",
-					"Prjn.PrjnScale.Rel": "0.5",
-				}},
-			{Sel: "#DGToCA3", Desc: "Mossy fibers: strong, non-learning",
-				Params: params.Params{
-					"Prjn.CHL.Hebb":      "0.001",
-					"Prjn.CHL.SAvgCor":   "1",
-					"Prjn.Learn.Learn":   "false",
-					"Prjn.WtInit.Mean":   "0.9",
-					"Prjn.WtInit.Var":    "0.01",
-					"Prjn.PrjnScale.Rel": "8",
-				}},
-			{Sel: "#CA3ToCA3", Desc: "CA3 recurrent cons",
-				Params: params.Params{
-					"Prjn.CHL.Hebb":      "0.01",
-					"Prjn.CHL.SAvgCor":   "1",
-					"Prjn.PrjnScale.Rel": "2",
-				}},
-			{Sel: "#CA3ToCA1", Desc: "Schaffer collaterals -- slower, less hebb",
-				Params: params.Params{
-					"Prjn.CHL.Hebb":    "0.005",
-					"Prjn.CHL.SAvgCor": "0.4",
-					"Prjn.Learn.LRate": "0.1",
-				}},
-			{Sel: ".EC", Desc: "all EC layers: only pools, no layer-level",
-				Params: params.Params{
-					"Layer.Acts.Gbar.L":          ".1",
-					"Layer.Inhib.ActAvg.Nominal": "0.2",
-					"Layer.Inhib.Layer.On":       "false",
-					"Layer.Inhib.Pool.Gi":        "2.0",
-					"Layer.Inhib.Pool.On":        "true",
-				}},
-			{Sel: "#DG", Desc: "very sparse = high inibhition",
-				Params: params.Params{
-					"Layer.Inhib.ActAvg.Nominal": "0.01",
-					"Layer.Inhib.Layer.Gi":       "3.6",
-				}},
-			{Sel: "#CA3", Desc: "sparse = high inibhition",
-				Params: params.Params{
-					"Layer.Inhib.ActAvg.Nominal": "0.02",
-					"Layer.Inhib.Layer.Gi":       "2.8",
-				}},
-			{Sel: "#CA1", Desc: "CA1 only Pools",
-				Params: params.Params{
-					"Layer.Inhib.ActAvg.Nominal": "0.1",
-					"Layer.Inhib.Layer.On":       "false",
-					"Layer.Inhib.Pool.Gi":        "2.2",
-					"Layer.Inhib.Pool.On":        "true",
-				}},
-		},
-	}},
-	{Name: "NoCHL", Desc: "no learning in CHL main hip pathways -- for debugging auto-encoder", Sheets: params.Sheets{
-		"Network": &params.Sheet{
-			{Sel: ".HippoCHL", Desc: "no learning",
-				Params: params.Params{
-					"Prjn.Learn.LRate": "0",
-				}},
-		},
-		"Sim": &params.Sheet{},
-	}},
-}
