--- examples/hip_bench/config.toml
@@ -1,1 +1,0 @@
-includes = ["smallhip.toml"]
